refactored and added new conversation steps for the careconvers simulation, updated environment variables for testing, and made various adjustments to the backend and frontend code to support the new features and improvements.
