+++ backend/.env.test.example
@@ -1,0 +1,31 @@
+# ============================================
+# Backend test environment template
+# Copy to: backend/.env
+# ============================================
+
+# Runtime
+NODE_ENV=development
+PORT=3000
+
+# Supabase (required)
+# Use your TEST Supabase project values
+SUPABASE_URL=https://your-test-project.supabase.co
+SUPABASE_KEY=your_test_supabase_anon_key
+SUPABASE_SERVICE_ROLE_KEY=your_test_supabase_service_role_key
+
+# Compatibility (optional aliases used in code)
+VITE_SUPABASE_URL=https://your-test-project.supabase.co
+VITE_SUPABASE_KEY=your_test_supabase_anon_key
+
+# CareConvers AI intent classification (optional)
+# If empty/missing, backend falls back to regex intent matching
+GEMINI_API_KEY=your_gemini_api_key_here
+
+# Optional CORS override for push route
+# Example: http://localhost:5173,http://127.0.0.1:5173
+ALLOWED_ORIGINS=
+
+# Optional Web Push config (only if testing push notifications)
+VAPID_PUBLIC=
+VAPID_PRIVATE=
+ADMIN_PUSH_KEY=
