+++ backend/supabase/careconvers/quizData.js
@@ -1,0 +1,126 @@
+const quizQuestions = [
+  {
+    id: 1,
+    type: 'MCQ',
+    question: "Comment peut-on aider une personne atteinte de la maladie d'Alzheimer à mieux comprendre?",
+    correct: 4,
+    options: [
+      "En évitant de lui parler pour ne pas la fatiguer.",
+      "En lui posant des questions ouvertes",
+      "En lui parlant rapidement pour qu'elle comprenne vite.",
+      "En utilisant des phrases courtes et des termes simples."
+    ]
+  },
+  {
+    id: 2,
+    type: 'MCQ',
+    question: "Quelle est une bonne pratique lors de la communication avec la personne présentant des troubles cognitifs ?",
+    correct: 4,
+    options: [
+      "Ne pas répéter les informations si elle ne comprend pas.",
+      "Se mettre de côté et lui parler proche de son oreille",
+      "Utiliser des questions ouvertes pour encourager la discussion.",
+      "Se mettre face à elle pour être dans son champ de vision."
+    ]
+  },
+  {
+    id: 3,
+    type: 'MCQ',
+    question: "Que faire si la personne atteinte de la maladie d'Alzheimer ne répond pas à votre présence?",
+    correct: 2,
+    options: [
+      "Parler plus fort",
+      "Attirer son attention en effleurant son épaule ou sa main.",
+      "Taper dans ses mains pour qu'elle réagisse.",
+      "L'ignorer jusqu'à ce qu'elle réponde."
+    ]
+  },
+  {
+    id: 4,
+    type: 'MCQ',
+    question: "Comment doit-on adapter l'environnement pour faciliter la communication?",
+    correct: 4,
+    options: [
+      "Mettre beaucoup de bruit pour stimuler la personne.",
+      "Allumer la télévision",
+      "Allumer toutes les lumières",
+      "Éviter la musique de fond et les lieux bruyants."
+    ]
+  },
+  {
+    id: 5,
+    type: 'MCQ',
+    question: "Quelle est une conséquence des troubles du langage chez la personne présentant des troubles cognitifs ?",
+    correct: 4,
+    options: [
+      "Elle apprécie beaucoup plus de parler.",
+      "Elle n'a plus besoin d'aide pour communiquer.",
+      "Elle est plus attentive aux discussions de son entourage.",
+      "Elle peut devenir incohérente dans ses discours."
+    ]
+  },
+  {
+    id: 6,
+    type: 'MCQ',
+    question: "Comment la communication non verbale peut-elle aider dans l'échange avec la personne présentant des troubles cognitifs ?",
+    correct: 2,
+    options: [
+      "Elle complique la compréhension.",
+      "Elle permet de transmettre des messages malgré la diminution de la parole.",
+      "Elle doit être évitée pour ne pas distraire la personne.",
+      "Elle n'a pas d'importance dans la communication."
+    ]
+  },
+  {
+    id: 7,
+    type: 'MCQ',
+    question: "Que doit-on faire si la personne atteinte de la maladie d'Alzheimer refuse de parler?",
+    correct: 2,
+    options: [
+      "Inciter la personne à parler, afin d'obtenir des réponses et ainsi éviter une perte du langage.",
+      "Respecter son silence et continuer à lui parler doucement.",
+      "Parler plus fort pour qu'elle réponde.",
+      "Lui poser plus de questions pour la faire parler."
+    ]
+  },
+  {
+    id: 8,
+    type: 'MCQ',
+    question: "Pourquoi est-il important de faire preuve de patience lors de la communication avec la personne présentant des troubles de la mémoire ?",
+    correct: 2,
+    options: [
+      "Parce qu'elle ne veut pas communiquer.",
+      "Parce qu'elle a souvent besoin de plus de temps pour comprendre et répondre.",
+      "Parce qu'elle doit apprendre à parler plus vite pour entretenir sa capacité à interagir.",
+      "Parce qu'elle doit être régulièrement encourgée à répondre rapidement."
+    ]
+  },
+  {
+    id: 9,
+    type: 'MCQ',
+    question: "Que doit-on faire si la personne malade exprime un comportement inapproprié?",
+    correct: 1,
+    options: [
+      "Ignorer ses propos et ne pas relever ses insultes.",
+      "Lui demander de se calmer immédiatement.",
+      "Réagir violemment pour lui faire comprendre.",
+      "Répondre en utilisant des propos vulgaires."
+    ]
+  },
+  {
+    id: 10,
+    type: 'MCQ',
+    question: "Une des règles pour communiquer efficacement avec une personne atteinte de démence est de :",
+    correct: 1,
+    options: [
+      "Parler clairement et se présenter",
+      "Parler rapidement pour capter l'attention",
+      "Ne pas utiliser de gestes",
+      "Éviter de lui poser des questions"
+    ]
+  }
+];
+
+module.exports = {
+  quizQuestions,
+};
