+++ backend/supabase/careconvers/utils.js
@@ -1,0 +1,17 @@
+function normalizeUserId(userId) {
+  if (typeof userId !== 'string') return '';
+  return userId.trim();
+}
+
+function shuffleArray(array) {
+  for (let i = array.length - 1; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+    [array[i], array[j]] = [array[j], array[i]];
+  }
+  return array;
+}
+
+module.exports = {
+  normalizeUserId,
+  shuffleArray,
+};
