+++ backend/supabaseClient.js
@@ -3,9 +3,23 @@ const path = require('path');
  
 // Essayer de charger le .env depuis le dossier upload ou utiliser les variables directement
 require('dotenv').config({ path: path.resolve(__dirname, '.env') });
- 
-const supabaseUrl = process.env.VITE_SUPABASE_URL || process.env.SUPABASE_URL;
-const supabaseKey = process.env.VITE_SUPABASE_KEY || process.env.SUPABASE_KEY;
+
+function isPlaceholder(value) {
+  const v = String(value || '').trim().toLowerCase();
+  return !v || v.includes('your-test-project') || v.includes('your_test_supabase');
+}
+
+const envSupabaseUrl = process.env.SUPABASE_URL;
+const viteSupabaseUrl = process.env.VITE_SUPABASE_URL;
+const supabaseUrl = !isPlaceholder(envSupabaseUrl)
+  ? envSupabaseUrl
+  : (!isPlaceholder(viteSupabaseUrl) ? viteSupabaseUrl : envSupabaseUrl || viteSupabaseUrl);
+
+const envSupabaseKey = process.env.SUPABASE_KEY;
+const viteSupabaseKey = process.env.VITE_SUPABASE_KEY;
+const supabaseKey = !isPlaceholder(envSupabaseKey)
+  ? envSupabaseKey
+  : (!isPlaceholder(viteSupabaseKey) ? viteSupabaseKey : envSupabaseKey || viteSupabaseKey);
 const supabaseServiceKey = process.env.SUPABASE_SERVICE_ROLE_KEY; // Clé de service pour bypass RLS
 const baseClientKey = supabaseKey || supabaseServiceKey;
 
@@ -43,13 +57,18 @@ if (supabaseServiceKey) {
   console.log('🔑 [DEBUG] Service key starts with:', supabaseServiceKey.substring(0, 10));
   
   // Décoder le JWT pour vérifier le rôle
-  try {
-    const payload = JSON.parse(Buffer.from(supabaseServiceKey.split('.')[1], 'base64').toString());
-    console.log('🔑 [DEBUG] JWT payload role:', payload.role);
-    console.log('🔑 [DEBUG] JWT payload iss:', payload.iss);
-    console.log('🔑 [DEBUG] Is service_role?', payload.role === 'service_role');
-  } catch (e) {
-    console.error('❌ [ERROR] Failed to decode JWT:', e.message);
+  const jwtParts = supabaseServiceKey.split('.');
+  if (jwtParts.length >= 2 && jwtParts[1]) {
+    try {
+      const payload = JSON.parse(Buffer.from(jwtParts[1], 'base64').toString());
+      console.log('🔑 [DEBUG] JWT payload role:', payload.role);
+      console.log('🔑 [DEBUG] JWT payload iss:', payload.iss);
+      console.log('🔑 [DEBUG] Is service_role?', payload.role === 'service_role');
+    } catch (e) {
+      console.warn('⚠️ [WARN] Service key not decodable as JWT (likely placeholder in local .env).');
+    }
+  } else {
+    console.warn('⚠️ [WARN] Service key format does not look like JWT (likely placeholder in local .env).');
   }
 } else {
   console.error('❌ [ERROR] Service key is missing! RLS bypass will not work.');
