+++ src/components.d.ts
@@ -7,8 +7,6 @@ export {}
 /* prettier-ignore */
 declare module 'vue' {
   export interface GlobalComponents {
-    Accordion: typeof import('primevue/accordion')['default']
-    AccordionTab: typeof import('primevue/accordiontab')['default']
     AchievementNotification: typeof import('./components/gamification/AchievementNotification.vue')['default']
     AchievementsPage: typeof import('./components/gamification/AchievementsPage.vue')['default']
     AddFileDocPFP: typeof import('./components/home/AddFileDocPFP.vue')['default']
@@ -30,10 +28,6 @@ declare module 'vue' {
     AppSkeleton: typeof import('./components/common/feedback/AppSkeleton.vue')['default']
     AppSpinner: typeof import('./components/common/feedback/AppSpinner.vue')['default']
     AuthForm: typeof import('./components/common/forms/AuthForm.vue')['default']
-    AutoComplete: typeof import('primevue/autocomplete')['default']
-    Avatar: typeof import('primevue/avatar')['default']
-    AvatarGroup: typeof import('primevue/avatargroup')['default']
-    Badge: typeof import('primevue/badge')['default']
     BadgeCard: typeof import('./components/gamification/BadgeCard.vue')['default']
     BadgeManagement: typeof import('./components/admin/BadgeManagement.vue')['default']
     BandeauGameMaster: typeof import('./components/gamification/BandeauGameMaster.vue')['default']
@@ -41,24 +35,16 @@ declare module 'vue' {
     BarChart: typeof import('./components/admin/widgets/charts/BarChart.vue')['default']
     BlockViewer: typeof import('./components/ui/BlockViewer.vue')['default']
     BracketView: typeof import('./components/tournois/BracketView.vue')['default']
-    Breadcrumb: typeof import('primevue/breadcrumb')['default']
-    Button: typeof import('primevue/button')['default']
     ButtonNavbar: typeof import('./components/ui/buttons/ButtonNavbar.vue')['default']
-    Calendar: typeof import('primevue/calendar')['default']
     CapsulesInsuffisanceRenaleNav: typeof import('./components/capsulesInsuffiance/CapsulesInsuffisanceRenaleNav.vue')['default']
     CapsulesInsuffisanceRenalePageToc: typeof import('./components/capsulesInsuffiance/CapsulesInsuffisanceRenalePageToc.vue')['default']
     CapsulesInsuffisanceRenaleTableDesMatieres: typeof import('./components/capsulesInsuffiance/CapsulesInsuffisanceRenaleTableDesMatieres.vue')['default']
-    Card: typeof import('primevue/card')['default']
     CardNameProfile: typeof import('./components/user/library/CardNameProfile.vue')['default']
     ChallengeCard: typeof import('./components/gamification/ChallengeCard.vue')['default']
     ChallengeManagement: typeof import('./components/admin/ChallengeManagement.vue')['default']
     ChallengesPage: typeof import('./components/gamification/ChallengesPage.vue')['default']
     ChartSelector: typeof import('./components/admin/widgets/ChartSelector.vue')['default']
     ChatProfil: typeof import('./components/user/details/ChatProfil.vue')['default']
-    Checkbox: typeof import('primevue/checkbox')['default']
-    Chip: typeof import('primevue/chip')['default']
-    ColorPicker: typeof import('primevue/colorpicker')['default']
-    Column: typeof import('primevue/column')['default']
     CommunitiesList: typeof import('./components/social/library/CommunitiesList.vue')['default']
     CommunityFeed: typeof import('./components/social/library/CommunityFeed.vue')['default']
     ConsigneModal: typeof import('./components/careconvers/ConsigneModal.vue')['default']
@@ -72,16 +58,10 @@ declare module 'vue' {
     DailyWheelModal: typeof import('./components/gamification/daily/DailyWheelModal.vue')['default']
     DailyWheelWidget: typeof import('./components/gamification/daily/DailyWheelWidget.vue')['default']
     DashboardKpiGrid: typeof import('./components/admin/widgets/DashboardKpiGrid.vue')['default']
-    DataTable: typeof import('primevue/datatable')['default']
-    DataView: typeof import('primevue/dataview')['default']
     DetailModal: typeof import('./components/gamification/DetailModal.vue')['default']
-    Dialog: typeof import('primevue/dialog')['default']
-    Divider: typeof import('primevue/divider')['default']
     DocumentsUserProfile: typeof import('./components/user/details/DocumentsUserProfile.vue')['default']
     DoughnutChart: typeof import('./components/admin/widgets/charts/DoughnutChart.vue')['default']
-    Dropdown: typeof import('primevue/dropdown')['default']
     EditFileDocPFP: typeof import('./components/home/EditFileDocPFP.vue')['default']
-    Editor: typeof import('primevue/editor')['default']
     EmptyState: typeof import('./components/common/feedback/EmptyState.vue')['default']
     EnseignentForm: typeof import('./components/admin/forms/EnseignentForm.vue')['default']
     EnseignentFormModif: typeof import('./components/admin/forms/EnseignentFormModif.vue')['default']
@@ -94,14 +74,12 @@ declare module 'vue' {
     EventDetail: typeof import('./components/events/EventDetail.vue')['default']
     EventForm: typeof import('./components/events/EventForm.vue')['default']
     FaqBox: typeof import('./components/admin/items/FaqBox.vue')['default']
-    FileUpload: typeof import('primevue/fileupload')['default']
     FilterComponent: typeof import('./components/social/library/FilterComponent.vue')['default']
     FilterInstitution: typeof import('./components/common/filters/FilterInstitution.vue')['default']
     FilterInstitution_SUPABASE: typeof import('./components/common/filters/FilterInstitution_SUPABASE.vue')['default']
     FilterSidebar: typeof import('./components/common/filters/FilterSidebar.vue')['default']
     FiltreMap: typeof import('./components/common/filters/FiltreMap.vue')['default']
     Footer: typeof import('./components/common/utils/Footer.vue')['default']
-    Galleria: typeof import('primevue/galleria')['default']
     GameHub: typeof import('./components/games/GameHub.vue')['default']
     GamificationActivityWidget: typeof import('./components/gamification/widgets/GamificationActivityWidget.vue')['default']
     GamificationNotification: typeof import('./components/gamification/notifications/GamificationNotification.vue')['default']
@@ -115,8 +93,6 @@ declare module 'vue' {
     HouseManagement: typeof import('./components/admin/HouseManagement.vue')['default']
     HousesRankingPage: typeof import('./components/gamification/HousesRankingPage.vue')['default']
     HouseStatsPage: typeof import('./components/gamification/HouseStatsPage.vue')['default']
-    IconField: typeof import('primevue/iconfield')['default']
-    Image: typeof import('primevue/image')['default']
     ImageCropper: typeof import('./components/social/library/ImageCropper.vue')['default']
     InfinityScroll: typeof import('./components/social/library/InfinityScroll.vue')['default']
     Info_repondant: typeof import('./components/admin/details/Info_repondant.vue')['default']
@@ -124,10 +100,6 @@ declare module 'vue' {
     InfoRepondantSignature: typeof import('./components/admin/details/InfoRepondantSignature.vue')['default']
     InformationUtilisation: typeof import('./components/home/InformationUtilisation.vue')['default']
     InformationView: typeof import('./components/home/InformationView.vue')['default']
-    InputIcon: typeof import('primevue/inputicon')['default']
-    InputNumber: typeof import('primevue/inputnumber')['default']
-    InputSwitch: typeof import('primevue/inputswitch')['default']
-    InputText: typeof import('primevue/inputtext')['default']
     InstitutionForm: typeof import('./components/admin/forms/InstitutionForm.vue')['default']
     InstitutionFormModif: typeof import('./components/admin/forms/InstitutionFormModif.vue')['default']
     InstitutionView: typeof import('./components/institutions/InstitutionView.vue')['default']
@@ -146,13 +118,9 @@ declare module 'vue' {
     Management_votation_prioritaire: typeof import('./components/admin/details/Management_votation_prioritaire.vue')['default']
     ManagementPFPEnCours: typeof import('./components/admin/details/ManagementPFPEnCours.vue')['default']
     ManagementPlacesSafe: typeof import('./components/admin/details/ManagementPlacesSafe.vue')['default']
-    Menu: typeof import('primevue/menu')['default']
-    Menubar: typeof import('primevue/menubar')['default']
-    Message: typeof import('primevue/message')['default']
     MiniChart: typeof import('./components/admin/widgets/MiniChart.vue')['default']
     MobileBottomNav: typeof import('./components/common/utils/MobileBottomNav.vue')['default']
     ModuleSection: typeof import('./components/video/ModuleSection.vue')['default']
-    MultiSelect: typeof import('primevue/multiselect')['default']
     Navbar: typeof import('./components/common/utils/Navbar.vue')['default']
     NewUserForm: typeof import('./components/admin/forms/NewUserForm.vue')['default']
     NewUserFormModif: typeof import('./components/admin/forms/NewUserFormModif.vue')['default']
@@ -162,11 +130,8 @@ declare module 'vue' {
     OffreDePlace: typeof import('./components/admin/details/OffreDePlace.vue')['default']
     OffreDePlaceBA24PFP2: typeof import('./components/admin/details/OffreDePlaceBA24PFP2.vue')['default']
     OverCreatePlace: typeof import('./components/admin/details/OverCreatePlace.vue')['default']
-    OverlayPanel: typeof import('primevue/overlaypanel')['default']
     PageHeader: typeof import('./components/admin/common/PageHeader.vue')['default']
     PageList: typeof import('./components/editor/notes/PageList.vue')['default']
-    Panel: typeof import('primevue/panel')['default']
-    Password: typeof import('primevue/password')['default']
     PdfViewerModal: typeof import('./components/careconvers/PdfViewerModal.vue')['default']
     PeriodComparisonPanel: typeof import('./components/admin/widgets/PeriodComparisonPanel.vue')['default']
     PeriodSelector: typeof import('./components/admin/widgets/PeriodSelector.vue')['default']
@@ -184,8 +149,6 @@ declare module 'vue' {
     ProfileAdminRightSidebar: typeof import('./components/user/library/ProfileAdminRightSidebar.vue')['default']
     ProfileInfo: typeof import('./components/user/details/ProfileInfo.vue')['default']
     ProfileInfoAdmin: typeof import('./components/user/details/ProfileInfoAdmin.vue')['default']
-    ProgressBar: typeof import('primevue/progressbar')['default']
-    ProgressSpinner: typeof import('primevue/progressspinner')['default']
     PublicCommunitiesList: typeof import('./components/social/library/PublicCommunitiesList.vue')['default']
     PushControls: typeof import('./components/push/PushControls.vue')['default']
     PwaInstallPrompt: typeof import('./components/common/utils/PwaInstallPrompt.vue')['default']
@@ -196,8 +159,6 @@ declare module 'vue' {
     QuestsProfileCard: typeof import('./components/gamification/QuestsProfileCard.vue')['default']
     QuestsSidebarCard: typeof import('./components/gamification/QuestsSidebarCard.vue')['default']
     RadarProfil: typeof import('./components/user/details/RadarProfil.vue')['default']
-    RadioButton: typeof import('primevue/radiobutton')['default']
-    Rating: typeof import('primevue/rating')['default']
     ResizableWidgetGrid: typeof import('./components/admin/widgets/ResizableWidgetGrid.vue')['default']
     ResumStageUserProfile: typeof import('./components/user/details/ResumStageUserProfile.vue')['default']
     RightSidebar: typeof import('./components/social/library/RightSidebar.vue')['default']
@@ -208,17 +169,12 @@ declare module 'vue' {
     ScenarioObjectivesModal: typeof import('./components/careconvers/ScenarioObjectivesModal.vue')['default']
     SearchButton: typeof import('./components/ui/buttons/SearchButton.vue')['default']
     SearchResults: typeof import('./components/common/utils/SearchResults.vue')['default']
-    SelectButton: typeof import('primevue/selectbutton')['default']
-    Sidebar: typeof import('primevue/sidebar')['default']
     SidebarMenuItems: typeof import('./components/admin/lists/SidebarMenuItems.vue')['default']
     SimpleTable: typeof import('./components/admin/widgets/charts/SimpleTable.vue')['default']
-    Skeleton: typeof import('primevue/skeleton')['default']
-    Slider: typeof import('primevue/slider')['default']
     SmartVisualization: typeof import('./components/admin/widgets/SmartVisualization.vue')['default']
     StageRepartitionBA2: typeof import('./components/admin/details/StageRepartitionBA2.vue')['default']
     StatsBox: typeof import('./components/admin/items/StatsBox.vue')['default']
     StatsCard: typeof import('./components/admin/widgets/StatsCard.vue')['default']
-    Steps: typeof import('primevue/steps')['default']
     StoriesBar: typeof import('./components/social/library/StoriesBar.vue')['default']
     StoryAudio: typeof import('./components/social/library/StoryAudio.vue')['default']
     StoryCarousel: typeof import('./components/social/library/StoryCarousel.vue')['default']
@@ -230,17 +186,12 @@ declare module 'vue' {
     StudentCreateDialog: typeof import('./components/admin/forms/StudentCreateDialog.vue')['default']
     StudentEditDialog: typeof import('./components/admin/forms/StudentEditDialog.vue')['default']
     SwitchColor: typeof import('./components/ui/buttons/SwitchColor.vue')['default']
-    TabPanel: typeof import('primevue/tabpanel')['default']
-    TabView: typeof import('primevue/tabview')['default']
-    Tag: typeof import('primevue/tag')['default']
     TeamSection: typeof import('./components/admin/items/TeamSection.vue')['default']
-    Textarea: typeof import('primevue/textarea')['default']
     TextAreaComponent: typeof import('./components/social/library/TextAreaComponent.vue')['default']
     TheNavbar: typeof import('./components/TheNavbar.vue')['default']
     TicketDetails: typeof import('./components/academic/TicketDetails.vue')['default']
     TicketForm: typeof import('./components/academic/TicketForm.vue')['default']
     TiptapSimpleEditor: typeof import('./components/editor/notes/TiptapSimpleEditor.vue')['default']
-    ToggleButton: typeof import('primevue/togglebutton')['default']
     TournoiDetails: typeof import('./components/tournois/TournoiDetails.vue')['default']
     TrackStatsWidget: typeof import('./components/admin/widgets/TrackStatsWidget.vue')['default']
     TriStateCell: typeof import('./components/admin/TriStateCell.vue')['default']
